ai_fitness_planner/requirements.txt
streamlit>=1.32.0
pandas
scikit-learn
Pillow
reportlab
huggingface_hub
transformers
requests
